+++ CHANGELOG.md
@@ -22,6 +22,8 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 | ---------------- | ------- | ------------------------------------------------------------------ |
 | Symbol Bootstrap | v1.0.3  | [symbol-bootstrap](https://www.npmjs.com/package/symbol-bootstrap) |
 
+-   Added `wizard` command.
+-   Added `pack` command.
 -   Improved Custom Preset Object types for symbol bootstrap lib integration.
 -   TransactionSelectionStrategy's new default value is `oldest`.